Extracted TaskStatus and TaskList as separated components
